requirements.txt
Django==1.9
djangorestframework==3.3.3
MySQL-python==1.2.5
slugify==0.0.1
wheel==0.24.0
